+++ flutter/lib/src/native/platform/media_devices.g.dart
@@ -1,5 +1,7 @@
 import 'dart:ffi';
+
 import 'package:ffi/ffi.dart';
+
 import 'package:medea_jason/src/native/ffi/foreign_value.dart';
 
 void registerFunction(
@@ -14,12 +16,14 @@ void registerFunction(
   required Pointer<NativeFunction<Handle Function()>> microphoneVolume,
   required Pointer<NativeFunction<Handle Function(Int64)>> setMicrophoneVolume,
   required Pointer<NativeFunction<Void Function(Handle)>> onDeviceChange,
+  required Pointer<NativeFunction<Int64 Function(Handle)>>
+      getMediaExceptionKind,
 }) {
   dl.lookupFunction<
       Void Function(Pointer, Pointer, Pointer, Pointer, Pointer, Pointer,
-          Pointer, Pointer),
+          Pointer, Pointer, Pointer),
       void Function(Pointer, Pointer, Pointer, Pointer, Pointer, Pointer,
-          Pointer, Pointer)>('register_media_devices')(
+          Pointer, Pointer, Pointer)>('register_media_devices')(
     enumerateDevices,
     getUserMedia,
     getDisplayMedia,
@@ -28,5 +32,6 @@ void registerFunction(
     microphoneVolume,
     setMicrophoneVolume,
     onDeviceChange,
+    getMediaExceptionKind,
   );
 }