+++ flutter/lib/src/web/room_handle.dart
@@ -84,13 +84,13 @@ class WebRoomHandle extends RoomHandle {
   }
 
   @override
-  Future<void> enableRemoteVideo() async {
-    await fallibleFuture(obj.enable_remote_video());
+  Future<void> enableRemoteVideo([MediaSourceKind? kind]) async {
+    await fallibleFuture(obj.enable_remote_video(kind?.index));
   }
 
   @override
-  Future<void> disableRemoteVideo() async {
-    await fallibleFuture(obj.disable_remote_video());
+  Future<void> disableRemoteVideo([MediaSourceKind? kind]) async {
+    await fallibleFuture(obj.disable_remote_video(kind?.index));
   }
 
   @override
@@ -123,7 +123,9 @@ class WebRoomHandle extends RoomHandle {
 
   @override
   void onFailedLocalMedia(void Function(Object) f) {
-    fallibleFunction(() => obj.on_failed_local_media(allowInterop(f)));
+    fallibleFunction(() => obj.on_failed_local_media(allowInterop((e) {
+          f(convertException(e));
+        })));
   }
 
   @moveSemantics