+++ src/api/dart/utils/mod.rs
@@ -16,7 +16,10 @@ pub use self::{
     arrays::PtrArray,
     err::{new_panic_error, ArgumentError, DartError},
     result::DartResult,
-    string::{c_str_into_string, free_dart_native_string, string_into_c_str},
+    string::{
+        c_str_into_string, dart_string_into_rust, free_dart_native_string,
+        string_into_c_str,
+    },
 };
 
 /// Rust representation of a Dart [`Future`].
@@ -65,3 +68,20 @@ where
         DartFuture(dart_future, PhantomData)
     }
 }
+
+/// Tries to convert the provided [`DartValueArg`] using [`TryInto`].
+///
+/// If the conversion fails, then [`ArgumentError`] is [`return`]ed as a
+/// [`DartFuture`].
+macro_rules! dart_arg_try_into {
+    ($k:expr) => {
+        match $k.try_into().map_err(|err: DartValueCastError| {
+            ArgumentError::new(err.value, "kind", err.expectation)
+        }) {
+            Ok(s) => s,
+            Err(e) => return async move { Err(e.into()) }.into_dart_future(),
+        }
+    };
+}
+
+pub(crate) use dart_arg_try_into;