+++ src/platform/dart/utils/callback.rs
@@ -165,6 +165,9 @@ impl Callback {
     /// Returns a [`Callback`] wrapping the provided [`FnMut`] with two
     /// arguments, that can be converted to a [`Dart_Handle`] and passed to
     /// Dart.
+    // TODO: False positive: complex bounds.
+    //       https://github.com/rust-lang/rust-clippy/issues/9076
+    #[allow(clippy::trait_duplication_in_bounds)]
     pub fn from_two_arg_fn_mut<F, T, S>(mut f: F) -> Self
     where
         F: FnMut(T, S) + 'static,
@@ -299,7 +302,7 @@ pub mod tests {
     pub unsafe extern "C" fn test_callback_listener_dart_handle() -> Dart_Handle
     {
         Callback::from_once(move |val: Dart_Handle| {
-            unsafe { (TEST_CALLBACK_HANDLE_FUNCTION.unwrap())(val) };
+            (TEST_CALLBACK_HANDLE_FUNCTION.unwrap())(val);
         })
         .into_dart()
     }